+++ src/char/txt/maildb_txt.c
@@ -428,7 +428,7 @@ bool maildb_txt_delete(int char_id)
 
 		for(i = 0; i < m->store; i++) {
 			// ƒyƒbƒgíœ
-			if(md[i].item.card[0] == (short)0xff00)
+			if(md[i].item.card[0] == (int)0xff00)
 				petdb_delete(*((int *)(&md[i].item.card[1])));
 		}
